requirements.txt
babel
Django==4.0.5
django-addanother==2.1.0
django-ajax-selects==2.2.0
django-bootstrap-datepicker-plus==4.0.0
django-bower==5.2.0
django-compressor==3.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-libsass==0.9
django-phonenumber-field
django-postgres-extra==2.0.5
django-redis
django-select2==7.10.0
django-widget-tweaks==1.4.12
libsass==0.21.0
lxml==4.9.0
phonenumbers==8.12.49
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycountry==22.3.5
python-dateutil>=2.8.2,<3.0.0
pytz>=2022.1
PyYAML>=6.0,<7.0
regex>=2022.6.2
requests==2.28.0
toml==0.10.2
tomli==2.0.1
typing_extensions>=4.2.0
django-ses
